Problem: 'basic' module is too big
Solution: extract stack-related functionality into its own module
