+++ src/script/mod.rs
@@ -346,6 +346,8 @@ use lmdb;
 
 use pubsub;
 
+pub mod basic;
+pub mod stack;
 pub mod storage;
 pub mod timestamp_hlc;
 pub mod hash;
@@ -426,6 +428,7 @@ impl<'a> VM<'a> {
             inbox: receiver,
             sender: sender.clone(),
             modules: vec![Box::new(basic::Handler::new(publisher)),
+                          Box::new(stack::Handler::new()),
                           Box::new(storage::Handler::new(db_env, db)),
                           Box::new(hash::Handler::new()),
                           Box::new(timestamp_hlc::Handler::new()),
@@ -648,7 +651,6 @@ impl<'a> VM<'a> {
     }
 }
 
-pub mod basic;
 pub mod compose;
 
 #[cfg(test)]
